+++ builder/rules/create_quote.py
@@ -13,7 +13,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(21, 40)
     def rule_woman_21_50_car(self):
         logging.debug("Car Woman, 21 - 40")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.11)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.11
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.car
@@ -21,7 +21,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(41, 65)
     def rule_woman_41_65_car(self):
         logging.debug("Car Woman, 41 - 65")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.09)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.09
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.car
@@ -29,7 +29,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(66, 75)
     def rule_woman_66_75_car(self):
         logging.debug("Car Woman, 66 - 75")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.155)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.155
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.car
@@ -37,7 +37,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(25, 35)
     def rule_man_25_35_car(self):
         logging.debug("Car Man, 25 - 35")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.15)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.15
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.car
@@ -45,7 +45,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(36, 60)
     def rule_man_36_60_car(self):
         logging.debug("Car Man, 36 - 60")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.12)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.12
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.car
@@ -53,7 +53,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(61, 75)
     def rule_man_61_75_car(self):
         logging.debug("Car Man, 61 - 75")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.135)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.135
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.moto
@@ -61,7 +61,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(21, 40)
     def rule_woman_21_50_moto(self):
         logging.debug("Moto Woman, 21 - 40")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.25)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.25
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.moto
@@ -69,7 +69,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(41, 65)
     def rule_woman_41_65_moto(self):
         logging.debug("Moto Woman, 41 - 65")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.23)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.23
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.moto
@@ -77,7 +77,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(66, 75)
     def rule_woman_66_75_moto(self):
         logging.debug("Moto Woman, 66 - 75")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.295)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.295
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.moto
@@ -85,7 +85,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(25, 35)
     def rule_man_25_35_moto(self):
         logging.debug("Moto Man, 25 - 35")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.29)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.29
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.moto
@@ -93,7 +93,7 @@ class CreateQuote(Rule):
     @decorators.bracket_age(36, 60)
     def rule_man_36_60_moto(self):
         logging.debug("Moto Man, 36 - 60")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.26)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.26
         logging.debug("Quote starting amount: %d", self.quote.amount)
 
     @decorators.moto
@@ -101,5 +101,5 @@ class CreateQuote(Rule):
     @decorators.bracket_age(61, 75)
     def rule_man_61_75_moto(self):
         logging.debug("Moto Man, 61 - 75")
-        self.quote.amount = int(self.quote.vehicule.value * 100 * 0.275)
+        self.quote.amount = self.quote.vehicule.value * 100 * 0.275
         logging.debug("Quote starting amount: %d", self.quote.amount)
